+++ CHANGELOG.md
@@ -1,7 +1,20 @@
 # Changelog
 
+## [0.1.10] - 2025-09-04
 
-## [1.0.5] - 2025-08-31
+### Added
+-   Added DaisyWhite4LevelLight, replacing DaisyWhiteLight
+
+### Removed
+-   Removed DaisyAwningsCover, it was replaced by DaisyShadeCover
+
+### Fixed
+- should fix https://github.com/andreasnuesslein/py-teleco-daisy/issues/11, https://github.com/andreasnuesslein/py-teleco-daisy/issues/12,
+- should also fix #5
+
+
+
+## [0.1.5] - 2025-08-31
 
 ### Added
 -   Added a 4CH Heater `dType 21, dModel 20` #23